% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airlines.R
\docType{data}
\name{airlines}
\alias{airlines}
\title{Airline names.}
\format{
A data frame with columns:
\describe{
\item{carrier}{Two letter abbreviation.}
\item{name}{Full name.}
}
}
\source{
\url{https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236}
}
\description{
Look up airline names from their carrier codes.
}
\keyword{datasets}
