% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcflights21.R
\docType{data}
\name{dcflights}
\alias{dcflights}
\title{Flights data}
\format{
A data frame with columns:
\describe{
\item{airline}{Airline name.}
\item{flight}{Flight number.}
\item{destination}{Destination airports (only the top 10 destination
airports are included in this dataset).}
\item{date}{Date of departure.}
\item{hour}{Time of scheduled departure by hour}
\item{visibility}{Visibility, in miles.}
\item{duration}{Scheduled amount of time spent in the air, in minutes.}
\item{dep_delay}{Departure delay, in
minutes. Negative values represent early departures/arrivals.}
}
}
\source{
RITA, Bureau of transportation statistics,
\url{https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236}

\url{https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236}

ASOS download from Iowa Environmental Mesonet,
\url{https://mesonet.agron.iastate.edu/request/download.phtml}.

\url{https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236}

FAA Aircraft registry,
\url{http://www.faa.gov/licenses_certificates/aircraft_certification/aircraft_registry/releasable_aircraft_download/}
}
\description{
Data for flights that departed from the Washington DC airport (Reagan
Internation) during July 2021 and arrived at one of the top 10 connecting
airports.
}
\keyword{datasets}
