% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcflights21.R
\docType{data}
\name{dcflights21}
\alias{dcflights21}
\alias{flights}
\title{Flights data}
\format{
A data frame with columns:
\describe{
\item{airline}{Airline name.}
\item{flight}{Flight number.}
\item{origin, destination}{Origin and destination airports.}
\item{date}{Date of departure.}
\item{hour}{Time of scheduled departure by hour}
\item{dep_delay}{Departure delay, in
minutes. Negative values represent early departures/arrivals.}
\item{plane}{Plane tail number, giving a unique ID for each plane.}
\item{air_time}{Scheduled amount of time spent in the air, in minutes.}
\item{duration}{Distance between airports, in miles.}
\item{wind_speed}{Wind speed, in miles per hour.}
\item{visibility}{Visibility, in miles.}
\item{year}{Year the flown plane was manufactured.}
}
}
\source{
RITA, Bureau of transportation statistics,
\url{https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236}

\url{https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236}

ASOS download from Iowa Environmental Mesonet,
\url{https://mesonet.agron.iastate.edu/request/download.phtml}.

\url{https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236}

FAA Aircraft registry,
\url{http://www.faa.gov/licenses_certificates/aircraft_certification/aircraft_registry/releasable_aircraft_download/}
}
\description{
Data for all flights that departed from airports serving the
Washington, DC area in 2021.
}
\keyword{datasets}
